New laptop; new tweaks.
* More README content.
* Adjust prompt colors and tmux colors.
* Check for .gitconfig before rm'ing in gen-git-config script.

+++ tools/util/gen-git-config.sh
@@ -1,7 +1,7 @@
 #!/bin/sh
 
 # Get rid of what's already there.
-rm ~/.gitconfig
+[ -f ~/.gitconfig ] && rm ~/.gitconfig
 
 # Generate the global git config file.
 git config --global push.default "simple"
